Add env map and tone mapping for new >4 bone skinning examples
Directional light was also removed from the "many bones" example
because it wasn't needed.

The env map would make more of a difference if the material had lower roughness
but it looks too bright when extra lights are added like in the "performance"
example even though tone mapping is on.