Sprite: take FOV into account if sizeAttenuation is false
Make `Sprite`s which use a `SpriteMaterial` with `sizeAttenuation` set to
`false` also consider the FOV of a Perspective Camera. This ensures
that `Sprite`s maintain a constant size in screen-space regardless of the
changes in the FOV , not only the distance from the camera.

The existing issue was that `Sprite`s would still change in size when
the FOV of the camera was modified, even if `sizeAttenuation` was set to
`false`.

This is particularly useful when animating the FOV of the camera if one
expects the Sprites with `sizeAttenuation` set to `false` to maintain
their original size.

To do this, modify the `sprite.glsl.js` shader to also take into
account the camera FOV.
Specifically, divide the `scale` factor by the tangent of half of the
FOV. This value is located at the [1][1] position in the perspective
projection matrix in OpenGL, so we use it directly.
See https://stackoverflow.com/a/57822509 for details.

Also add an example that shows non-size attenuated `Sprite`s
specifically, with sliders that change the distance of the camera and
the FOV of the camera.