Fix missing footnote
It looks like the previous footnote for the KHR_lights_punctual extension was to mention that useLegacyLights needed to be disabled. The footnote itself was removed in a previous commit, but the reference to it remained despite there being no "footnote 1" any longer.

Removed the footnote marker, and renumbered footnote 2 to footnote 1.