Create RenderTarget.js
This is a copy of the renderTarget from the three.module.js file with the necessary imports from three.module.js. WGSL requires a strict separation between texture read and texture write accessors. For the compute shader, texture node is used for read and textureStore node for write to separate the bindings. Since the StorageTexture class is located here, I also copied the RenderTarget here. At the moment the renderTarget does nothing other than the same as the renderTarget in three.module.js. This renderTarget can now be specifically expanded to avoid the error messages caused by the access of WGSL shaders (read access) and the renderTarget (write access) to the same resource.