XRHandModelFactory: Fix Inverted Hands
On the Apple Vision Pro, hand/controller indices are assigned by the order they appear, and unassigned when they disappear.

This can cause hands that initially appeared as one handedness to later appear as another handedness.   This causes the Mesh Hands to appear inverted, as the SkinnedMesh is contorted to the mirror bone structure.

This PR detects handedness flips and re-initializes those meshes when it occurs.