Fix GLTFExporter.js by checking if attenuation distance is set
Having no attenuation distance set in the project/code, defaults to infitity. This property then cannot be set in GLTF file and automatically sets to NULL. NULL property cannot be read by various softwares, for example - Blender 4 - > onwards.

By adding check if distance is !== infinity we either send that parameter if its set, or don't send it if its default.