Fix devtools littering window object (#30950)
The devtools Chrome extension injects `bridge.js` directly into the page content. As a result, any top-level definitions in `bridge.js` are set on `window`. This resulted in `window.dispatchEvent` being overwritten, which interfered with normal use of `window.dispatchEvent` by any site that relied on it.

This change wraps the entire `bridge.js` file in an anonymous function to prevent any littering of the `window` objcect.