Line2NodeMaterial: Remove unused `lineWidth` property. (#31401)
* Update Line2NodeMaterial.js

In the global environment of three.js, the `lineWidth` in `Line2NodeMaterial` does not take effect and is essentially ineffective. To modify the line width, it is necessary to use the `linewidth` setting. Additionally, neither `Material` nor `NodeMaterial` contains a `linewidth` property. Therefore, the `lineWidth` property in `Line2NodeMaterial` should be changed to `linewidth`, and its default value should be set to 1.

* Update Line2NodeMaterial.js

delete `linewidth` property.