Material: Fix copy method to include allowOverride property
The copy method in Material was missing the allowOverride property that is defined in the constructor. When cloning a Material instance, this property was not copied to the new material, causing incorrect behavior when allowOverride was set to false.

Fixed by ensuring allowOverride is properly copied in the copy method.