Fix: ShaderMaterial.copy() missing linewidth property
The copy method in ShaderMaterial was missing the linewidth property that is defined in the constructor. When cloning a ShaderMaterial instance, this property was not copied to the new material, causing incorrect behavior when linewidth is set to a value other than the default.

Fixed by ensuring linewidth is properly copied in the copy method.